requirements.txt
fastapi==0.115.1
uvicorn==0.27.1
sqlalchemy==2.0.40
asyncpg==0.29.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.7
alembic==1.13.1
httpx==0.27.0
slowapi==0.1.9
pydantic==2.6.1
python-dotenv==1.0.1
bcrypt==4.1.2
email-validator==2.1.0 